Fix broken HTML encoding
Missing to specify encoding can mess some clients which may break
the file encoding or render it incorrectly.

Example: file that loads ok locally, after upload to S3 gets broken
if the encoding is not mentioned in file.

Use of UTF-8 is a very good practice, but we still need to mention
it.